Don't Move to New Tag if You're Already Alone on the Last
So, if I had tags 1 2 and 3 filled and I was looking at the only
client on 3, and I did movenewtag, I would end up looking at 4.
And there'd be nothing there, the client stayed on 3.

That was no good.

So, I solved this by saying "Find me the next free tag not counting
the current client"

This way if the client isn't alone then the tag is still considered
occupied and it all works out.
If it is alone then it's current tag is considered the last one
and it's moved to the same place.
Win.