More Removing of Tags
So, if I was on the last tag and I last moved the item away, then it
would remain named.
I could navigate away and it wouldn't show up, but if I moved items
to a new tag, there it'd be again.
The name was risen!

Now I just say "If I'm not going to draw it, and it has a name, fix that"