Hacked Around SegFault in Click Handler
So, I made it so that the labels can be null.
The click handler, though, to figure out where the click landed was
checking how wide the labels were to see where the mouse was.

This was causing segfaults, though.
So, I hacked in a quick fix involving cut-and-paste.
It's not ideal, because it has no sense that most of the labels aren't
shown.
So, clicking in the middle of nowhere selects 27 or something.

Meh.
I don't use the feature, so I don't mind that it's broken.
I just don't want it to crash.